requirements.txt
asyncio==3.4.3
asyncio-nats-client==0.5.0